+++ app.py
@@ -1,0 +1,5 @@
+from src.bot import run_bot
+
+
+# discord runner
+run_bot()
