+++ src/__init__.py
@@ -1,0 +1,4 @@
+from dotenv import load_dotenv
+
+
+load_dotenv()
