+++ src/message.py
@@ -1,0 +1,2 @@
+async def get_pong(ctx):
+    await ctx.send("Pong!")
